update parameter icon
