+++ app/features/graph/components/canvas/common/index.ts
@@ -1,3 +1,5 @@
 import ElementContainer from './ElementContainer'
 
+export { ParameterIcon } from './ParameterIcon'
+
 export { ElementContainer }
