--- app/features/graph/components/icons/ParameterIcon/index.ts
@@ -1,1 +1,0 @@
-export { ParameterIcon } from './ParameterIcon'
